Fixing undefined references during linking
The use of "static constexpr" in SurfaceSplineBEZ.h, function
make_triangle_mesh(), causes problem with gcc v6.2.0, which generates
undefined references to two of the three defined static constexpr objects.
Removing the "static" keyword seems to fix the problem. The underlying cause
may be that by the standard, objects qualified with "static" have to be
accessible at runtime, which is why references are generated, but they are
nowhere defined. I am not sure, however, if the compiler should define these
symbols along with the code emitted when the templated function is used.