FIX: first attempt at fixing make_triangle_mesh()
"Fixing" here refers to adapt the code to the conventions for transition
maps I used for generating transition matrices. This patch reverses the
direction in which transition maps are applied: To find the position of
an unknown vertex, we perform a transition in the opposite direction to
find coordinates that express the position of our new vertex with
respect to a known frame.