Adding make_indices() methods
The separation of TriangleMesh formats into TriangleMesh and
TriangleGraph removed the common getIndices() method. The new
make_indices() methods, available for both TriangleMesh and
TriangleGraph, will build a *copy* of the triangle indices. This is
simple copy for TriangleMesh, whereas for TriangleGraph we reconstruct
the vertex indices from edge data.